# generated/backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DDF 精密五金 MES API",
    "version": "1.0.0",
    "description": "由 DomainDriveFramework 引擎自动生成",
    "contact": {
      "name": "DDF Engine"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "本地开发" }
  ],
  "paths": {
...
    "/api/inspection-plan": {
      "get": {
        "summary": "分页查询抽样方案",
        "tags": ["InspectionPlan"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/InspectionPlan" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建抽样方案",
        "tags": ["InspectionPlan"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InspectionPlanCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/inspection-plan/{id}": {
      "get": {
        "summary": "查询抽样方案详情",
        "tags": ["InspectionPlan"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新抽样方案",
        "tags": ["InspectionPlan"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InspectionPlanCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除抽样方案",
        "tags": ["InspectionPlan"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/production-line": {
      "get": {
        "summary": "分页查询产线",
        "tags": ["ProductionLine"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/ProductionLine" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建产线",
        "tags": ["ProductionLine"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductionLineCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/production-line/{id}": {
      "get": {
        "summary": "查询产线详情",
        "tags": ["ProductionLine"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新产线",
        "tags": ["ProductionLine"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductionLineCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除产线",
        "tags": ["ProductionLine"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/process-step": {
      "get": {
        "summary": "分页查询工序",
        "tags": ["ProcessStep"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/ProcessStep" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建工序",
        "tags": ["ProcessStep"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProcessStepCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/process-step/{id}": {
      "get": {
        "summary": "查询工序详情",
        "tags": ["ProcessStep"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新工序",
        "tags": ["ProcessStep"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProcessStepCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除工序",
        "tags": ["ProcessStep"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
    "/api/quality-category": {
      "get": {
        "summary": "分页查询质量分类",
        "tags": ["QualityCategory"],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" } },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/QualityCategory" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建质量分类",
        "tags": ["QualityCategory"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QualityCategoryCreate" }
            }
          }
        },
        "responses": {
          "201": { "description": "创建成功" }
        }
      }
    },
    "/api/quality-category/{id}": {
      "get": {
        "summary": "查询质量分类详情",
        "tags": ["QualityCategory"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "实体详情" }
        }
      },
      "put": {
        "summary": "更新质量分类",
        "tags": ["QualityCategory"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QualityCategoryCreate" }
            }
          }
        },
        "responses": {
          "200": { "description": "更新成功" }
        }
      },
      "delete": {
        "summary": "删除质量分类",
        "tags": ["QualityCategory"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "删除成功" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InspectionPlan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "plan_code": {
            "type": "string"
          },
          "plan_name": {
            "type": "string"
          },
          "check_type": {
            "type": "string"
          },
          "batch_min": {
            "type": "integer"
          },
          "batch_max": {
            "type": "integer"
          },
          "sample_size": {
            "type": "integer"
          },
          "accept_limit": {
            "type": "integer",
            "description": "允许的最大不合格数"
          },
          "reject_limit": {
            "type": "integer",
            "description": "拒收的最小不合格数"
          },
          "severity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        }
      },
      "InspectionPlanCreate": {
        "type": "object",
        "properties": {
          "plan_code": {
            "type": "string",
            "nullable": false
          },
          "plan_name": {
            "type": "string",
            "nullable": false
          },
          "check_type": {
            "type": "string",
            "nullable": false
          },
          "batch_min": {
            "type": "integer",
            "nullable": false
          },
          "batch_max": {
            "type": "integer",
            "nullable": false
          },
          "sample_size": {
            "type": "integer",
            "nullable": false
          },
          "accept_limit": {
            "type": "integer",
            "nullable": false
          },
          "reject_limit": {
            "type": "integer",
            "nullable": false
          },
          "severity": {
            "type": "string",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
      "ProductionLine": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "workshop_id": {
            "type": "integer"
          },
          "line_code": {
            "type": "string"
          },
          "line_name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ProductionLineCreate": {
        "type": "object",
        "properties": {
          "workshop_id": {
            "type": "integer",
            "nullable": false
          },
          "line_code": {
            "type": "string",
            "nullable": false
          },
          "line_name": {
            "type": "string",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
      "ProcessStep": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "process_route_id": {
            "type": "integer"
          },
          "step_no": {
            "type": "integer"
          },
          "step_code": {
            "type": "string"
          },
          "step_name": {
            "type": "string"
          },
          "standard_time_minutes": {
            "type": "integer"
          },
          "device_type_required": {
            "type": "string"
          },
          "workstation_type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "ProcessStepCreate": {
        "type": "object",
        "properties": {
          "process_route_id": {
            "type": "integer",
            "nullable": false
          },
          "step_no": {
            "type": "integer",
            "nullable": false
          },
          "step_code": {
            "type": "string",
            "nullable": false
          },
          "step_name": {
            "type": "string",
            "nullable": false
          },
          "standard_time_minutes": {
            "type": "integer",
            "nullable": false
          },
          "device_type_required": {
            "type": "string",
            "nullable": false
          },
          "workstation_type": {
            "type": "string",
            "nullable": false
          },
          "description": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
      "QualityCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "category_code": {
            "type": "string"
          },
          "category_name": {
            "type": "string"
          },
          "level": {
            "type": "integer"
          },
          "parent_id": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "QualityCategoryCreate": {
        "type": "object",
        "properties": {
          "category_code": {
            "type": "string",
            "nullable": false
          },
          "category_name": {
            "type": "string",
            "nullable": false
          },
          "level": {
            "type": "integer",
            "nullable": false
          },
          "parent_id": {
            "type": "integer",
            "nullable": false
          },
          "status": {
            "type": "string",
            "nullable": false
          }
        }
      },
...
    }
  }
}