# server/src/routes/inspection-plan.routes.js
/**
 * InspectionPlan CRUD 路由
 * Domain: base.inspection_plan
 * Generated by DDF Engine
 */

const express = require('express');
const router = express.Router();
const prisma = require('../prisma/client');

// 字段白名单（排除审计字段）
const FIELDS = ['plan_code', 'plan_name', 'check_type', 'batch_min', 'batch_max', 'sample_size', 'accept_limit', 'reject_limit', 'severity', 'status'];

// 可排序字段列表（由 DDF Engine 在生成期计算）
const SORTABLE_FIELDS = ['id', 'plan_code', 'plan_name', 'batch_min', 'batch_max', 'sample_size', 'accept_limit', 'reject_limit', 'created_at'];

// 可搜索字段列表（由 DDF Engine 在生成期计算）
const SEARCHABLE_FIELDS = ['plan_code', 'plan_name', 'check_type', 'severity', 'status'];

// ---- 字段值映射函数 ----
function mapInput(body) {
  const data = {};
  for (const key of FIELDS) {
    if (body[key] !== undefined) {
      data[key] = body[key];
    }
  }
  return data;
}

/**
 * GET /api/inspection-plan
 * 列表查询（分页 + 排序 + 搜索）
 */
router.get('/', async (req, res) => {
  try {
    const { page = 1, pageSize = 20, sortBy, sortOrder, search, ...filters } = req.query;
    const skip = (parseInt(page) - 1) * parseInt(pageSize);
    const take = parseInt(pageSize);

    const where = { ...filters };

    // 全文搜索
    if (search) {
      where.OR = SEARCHABLE_FIELDS.map(field => ({
        [field]: { contains: search },
      }));
    }

    // 排序（仅允许在 SORTABLE_FIELDS 范围内）
    let orderBy = { id: 'desc' };
    if (sortBy && SORTABLE_FIELDS.includes(sortBy)) {
      orderBy = { [sortBy]: sortOrder || 'asc' };
    }

    const [data, total] = await Promise.all([
      prisma.inspectionPlan.findMany({ where, skip, take, orderBy }),
      prisma.inspectionPlan.count({ where }),
    ]);

    res.json({ data, total, page: parseInt(page), pageSize: parseInt(pageSize) });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

/**
 * GET /api/inspection-plan/:id
 * 详情
 */
router.get('/:id', async (req, res) => {
  try {
    const data = await prisma.inspectionPlan.findUnique({
      where: { id: parseInt(req.params.id) },
    });
    if (!data) return res.status(404).json({ error: 'Not found' });
    res.json(data);
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

/**
 * POST /api/inspection-plan
 * 创建
 */
router.post('/', async (req, res) => {
  try {
    const data = await prisma.inspectionPlan.create({
      data: mapInput(req.body),
    });
    res.status(201).json(data);
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

/**
 * PUT /api/inspection-plan/:id
 * 更新
 */
router.put('/:id', async (req, res) => {
  try {
    const data = await prisma.inspectionPlan.update({
      where: { id: parseInt(req.params.id) },
      data: mapInput(req.body),
    });
    res.json(data);
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

/**
 * DELETE /api/inspection-plan/:id
 * 删除
 */
router.delete('/:id', async (req, res) => {
  try {
    await prisma.inspectionPlan.delete({
      where: { id: parseInt(req.params.id) },
    });
    res.status(204).send();
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

module.exports = router;