# server/src/routes/process-step.routes.js
/**
 * ProcessStep CRUD 路由
 * Domain: base.process_step
 * Generated by DDF Engine
 */

const express = require('express');
const router = express.Router();
const prisma = require('../prisma/client');

// 字段白名单（排除审计字段）
const FIELDS = ['process_route_id', 'step_no', 'step_code', 'step_name', 'standard_time_minutes', 'device_type_required', 'workstation_type', 'description'];

// 可排序字段列表（由 DDF Engine 在生成期计算）
const SORTABLE_FIELDS = ['id', 'step_no', 'step_code', 'step_name', 'standard_time_minutes', 'workstation_type', 'description', 'created_at', 'updated_at'];

// 可搜索字段列表（由 DDF Engine 在生成期计算）
const SEARCHABLE_FIELDS = ['step_code', 'step_name', 'device_type_required', 'workstation_type', 'description'];

// ---- 字段值映射函数 ----
function mapInput(body) {
  const data = {};
  for (const key of FIELDS) {
    if (body[key] !== undefined) {
      data[key] = body[key];
    }
  }
  return data;
}

/**
 * GET /api/process-step
 * 列表查询（分页 + 排序 + 搜索）
 */
router.get('/', async (req, res) => {
  try {
    const { page = 1, pageSize = 20, sortBy, sortOrder, search, ...filters } = req.query;
    const skip = (parseInt(page) - 1) * parseInt(pageSize);
    const take = parseInt(pageSize);

    const where = { ...filters };

    // 全文搜索
    if (search) {
      where.OR = SEARCHABLE_FIELDS.map(field => ({
        [field]: { contains: search },
      }));
    }

    // 排序（仅允许在 SORTABLE_FIELDS 范围内）
    let orderBy = { id: 'desc' };
    if (sortBy && SORTABLE_FIELDS.includes(sortBy)) {
      orderBy = { [sortBy]: sortOrder || 'asc' };
    }

    const [data, total] = await Promise.all([
      prisma.processStep.findMany({ where, skip, take, orderBy }),
      prisma.processStep.count({ where }),
    ]);

    res.json({ data, total, page: parseInt(page), pageSize: parseInt(pageSize) });
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

/**
 * GET /api/process-step/:id
 * 详情
 */
router.get('/:id', async (req, res) => {
  try {
    const data = await prisma.processStep.findUnique({
      where: { id: parseInt(req.params.id) },
    });
    if (!data) return res.status(404).json({ error: 'Not found' });
    res.json(data);
  } catch (err) {
    res.status(500).json({ error: err.message });
  }
});

/**
 * POST /api/process-step
 * 创建
 */
router.post('/', async (req, res) => {
  try {
    const data = await prisma.processStep.create({
      data: mapInput(req.body),
    });
    res.status(201).json(data);
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

/**
 * PUT /api/process-step/:id
 * 更新
 */
router.put('/:id', async (req, res) => {
  try {
    const data = await prisma.processStep.update({
      where: { id: parseInt(req.params.id) },
      data: mapInput(req.body),
    });
    res.json(data);
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

/**
 * DELETE /api/process-step/:id
 * 删除
 */
router.delete('/:id', async (req, res) => {
  try {
    await prisma.processStep.delete({
      where: { id: parseInt(req.params.id) },
    });
    res.status(204).send();
  } catch (err) {
    res.status(400).json({ error: err.message });
  }
});

module.exports = router;